refactor static folder
