+++ rel_bball_analytics/static/css/player-details.css
