+++ rel_bball_analytics/templates/players/details.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 
 {% block styles %}
-<link rel="stylesheet" href="{{ url_for('static', filename='styles/player-details.css') }}">
+<link rel="stylesheet" href="{{ url_for('static', filename='css/player-details.css') }}">
 {% endblock %}
 
 {% block content %}
